ch09/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.3
pip==22.2.2
setuptools==65.5.0
sqlparse==0.4.3
tzdata==2022.6
wheel==0.37.1
wincertstore==0.2
pkgconfig==1.5.5
mysqlclient==2.2.0
Markdown==3.4.4
django-simple-captcha==0.5.18
pymongo==4.5.0
plotly==5.16.1
numpy==1.25.2